gh-525 add script which blocks until dev-importer completes
... and either succeed or timeout after a set amount of seconds.

As of now it's not running on CI yet, but works fine locally. Next up is
to try it out on CI. Then we'll have an easy way to tell from our CI
whether the development path is blocked or not. I.e. in the future we
should no longer merge any PRs which don't pass this test.