gh-602 register callback to network for when peers change
The callback is intended to be filled with the schema udpate manager,
this way we can easily rebuild the graphql schema, whenever a new peer
is added to the network.

The callback is only called when the peers have actually changed, to
avoid constantly rebuilding the graphql api even when the network is
constant, but we received a ping for another reason.