gh-602 validate peer (in first field of path) is an existing peer
Also removed some coupling between the graphqlapi and network packages.
Previously the graphql.Build() would be called with a []network.Peer,
but it doesn't really care about the internals of a peer, it really just
wants a list of names, so it is now simply a []string. This removes one
cross-import of the network package from inside the graphqlapi package.