gh-609: in acceptance test runner remove orphans on d-c down
We have two docker-compose files in this repo. So if you start up the
dev one and then the test one it won't be able to do a poper
docker-compose down, because some containers are running in the same
network that are not specified in the current compose file.

Docker-compose considers those containers "orphans". By default stopping
a network with orphans errors, as they might still be in use. We can
circumvent that by specifying the `--remove-orphans` flag which means
all orphaned containers in the current network wil be stopped as well.

This gives us the ability to create a clean slate before our test runs.