gh-609: fix acceptance tests which compare schema to prototype
They do an introspection query on Network.* for 'WeaviateB' and
'WeaviateC'. Since we now have an actual ability to have different
schemas per instance they were failing as B and C had no schema.

So this now adds another backend fake that provides the schema that is
desired by the test.