gh-658: on import validate (and accept) network refs
This runs the whole validation process including whether the network
peer exists and whether the class on that peer exists. This might lead
to some timing inssues in the acceptance tests because we don't know
whether the network will be registered yet, by the time the import runs.
So we could consider importing the class with the reference only when
the test has reported that the network is ready.

As of now the graphql tree creation still fails because it can't handle
the network class ref.