gh-658 fix broken acceptance test for actions
Prior to the the change from the previous commit we had never used a
patch with multiple refs (i.e. array type) because the importer only
supported a single ref.

In order to get it working we had to change one type assertion were
previous code assumed a pointer type, but on a multi-ref patch we got a
non-pointer type.

However, this broke the case of put where we indeed get a pointer type,
so now we have a very small logic to deref the pointer type only if needed.