gh-658: adhere to 'localhost' convention in local ref in acc test
As mentioned previously we are now enforcing the convention that a local
reference must always have 'localhost' as the location URL. Previously
it was also allowed to match the hostname under which the weaviate
instance would start up. So this means previosly 127.0.0.1 was also
allowed for a local reference.

We still had some test that used the exact host, i.e. 127.0.0.1. This
commit should replace all of them with localhost and fix the broken
tests.