gh-659: (refactor) split out graphql Local.Get.Kind.Class
These are the places I will have to touch to also allow a
Network-Cross-Ref in. The logic that we have here is relatively simple,
but it was quite hard to understand because methods were a hundred lines
long.

This commit splits out the seperate concerns into different methods (and
files). This will allow us to build the Network-Cross-Refs into graphql soon.