gh-630 resolve Network.Get where filters at new position
Basically this is just deleting a bunch of code. Previously we had to
parse the filters entirely, because we had to modify the path. It looked
somehting like

WeaviateB->Things->City->name and was placed at the level of Network.Get.

We had to rewrite this to just say Things->City->name (i.e. remove the
peer name) and inject the filter at the correct place after
Local.Get.

Now that the filters are at the level of the KindClass anyway, we just
proxy them together with everything else. So no more parsing is required
at all.