gh-660: try to improve error handling for graphql response
We seem to be doing some interesting type conversion, we're first
marshalling the entire graphql response to JSON and then unmarshalling
it again into the go-swagger resopnse type. This seems to work fine for
most cases, but there is an edge cases where it breaks miserably:

When the graphql error response contains a list the path array will be
something like ["Get", "Things", 0, "foobar"]. Notice the third element
is a number, whereas all the other ones are string. Therefore this type
is defined in the graphql package as []interface{}. Howevever, we are
using OpenAPI2 with go-swagger, where variable types ("oneOf", "anyOf")
are not supported yet. Therefore the code generated by go-swagger has
the path field as []string.

The above case breaks the marhshal->unmarshal->marshal chain. Previously
it would fail silently. This commit makes sure the failure is at least
not silent anymore. Nevertheless it's not a very nice response yet. We
should improve upon this by actually converting from one response type
to the other (we can then simply convert the int to a string), rather
than using this marshal->unmarshal-marhsal magic. (#669)