gh-660: no longer return thunk in connector.LocalGetMeta
The original intend of the thunks was to enable parallelism. We want to
keep this idea, but the thunk can always be created in the
graphql.Resolve() methode in the graphqlapi package. The connectors do
not have to know of the concept of thunks, their concern should simply
be to resolve based on the given input.