gh-653: spike out happy path with hard-coded args
For now we are skipping the actual extraction from graphql, but rather
just hard-code the desired prop. In the connector we are hard-coding the
query for one specific type.

This is meant as a proof of concept and to discover what areas need to
be changed to thoroughly implement this feature (for all cases). From
the findings of this spike we can now successfully build a todo list of
what needs to be done throughout the completion of this issue.