gh-653: add totalTrue prop, introduce meta processor
The janusgraph/meta subpackage now has two components:

- The Query Builder contains everything that happens before we interact
with Janus, i.e. it uses our parsed arguments and creates a query based
on the desired props. The query is modelled to return the closest
possible result that we need for the graphQL API. Where an exact match
between our requirements for the graphQL API and what Janus returns is
not possible, the next component comes into place:

- The Query Processor sends the Query to Janus and then post-processes
the results where required. It merges props, where Janus decides to
split a map with multiplie keys into an array back into a map.
Additionally it can do arbitrary post processing per type.

This commit also adds the "totalTrue" prop to boolean props. In Janus
the counting of true/false props is one atomic unit, so we don't
seperate by counting true and false. I.e. whenever you ask for one of
the two, we also retrieve the other one. This in turn gives us the
safety to know that both will always be present. We can thus use the
post-processing step to calculate the percentages as well. Something
that I couldn't figure out how to do it Janus directly. However, this
computation is never expensive, so that shouldn't be an issue.

The atomic retrieval of both the true and false prop does however mean
that with the current logic of the query builder if both props are
specified we would retrieve both props twice (or even four times if all
four of totalTrue, totalFalse, percentageTrue, percentageFalse are set).
The next commit will include a small logic to prevent thus. Thus as of
now any prop other than the 'totalTrue' will error.