gh-653: support int props: calculate count
Previoulsly we pretended that every prop would always be a bool, this
enabled us to start quickly and not have to worry about analyzing the
type of the individual property.

Since we have now added the second type (int) this switching was
required and thus added. So far we only support "count" as a statistical
analysis prop, but have a good base to add others as well.

Please note that the way we aggregate ints leads to a special case when
only one prop is requested:

  just one analysis prop is a special case, because in multiple cases we
  are using select(<1>,<2>, ...<n>), this means we will receive a map
  that has the selections as keys. However, if we only ask for a single
  prop, Gremlin doesn't see a need to return a map and simply returns
  the primitive value. This will of course either break our
  post-processing or will not have the format the graphql API expects.
  We thus need to add an additional as().project().by() step to wrap the
  primtive prop in a map - but only if it's only a single analysis prop.