gh-653: extrapolate missing bool data if all are of the same value
After introducing filter (see previous commit message for details) we
found out that if we have only boolean props which are either all true
or all false, we miss data about the opposite value (which is zero in
this case). This commit makes sure we don't error, but instead
extrapolate the data we're missing.