gh-653: add acceptance test for many props
This test highly depends on a stable data set during test situations. So
far this data set has never changed, so this would have not been a
problem in the past. If we find out that we need to change the test data
set more often in the future, we have to find a better solution for this
test. Until then, this should elegantly solve the taks of verifying that
the GetMeta feature works and works correctly.