gh-646: introduce AssertEventuallyEqual() to test for delayed responses
It seems after removal of the keys the requests have become so fast that
the async behavior of janus and cassandra is not always ready yet,
leading to flakyness.

The introduced helper function retries every 10ms if the result is not
successful.

This should fix all the previously flaky "update action" tests, but it
might be that we now discover other areas to suffer from the same
problem. Let's see :)