gh-644: add getmeta resolver to graphqlapi network package
Independently both components work, but they don't play together yet,
because some hihgher up resolvers are still stubbed out. To fix this
(properly) I wanted to introduce a component test. This test would test
the entire network/getmeta package. However, I then found out that the
getmeta package isn't as well isolated as I though it was. It leaks
quite a bit into to the network overall package. We have to clean this
up first, and then we can continue with with building getmeta as
originally intended.

So expect a major refactor commit comming up soon.