gh-688: re-generate swagger code and fix unit test
A unit test in the network package was affected. Previously it seems it
changed the payload of the hash, however we accepted the changed value
in the unit test. I can't really see what changed, but the way the unit
test is now is actually more correct than before, so all is good.