gh-685: replace local RWMutex (schema, connector lock) with etcd RWMutex
Generally the prepared interface (RWMutex) had to be implemted. There
were a few changes to it. Since the Lock is now a remote lock, we need
a remote connection. This in turn means that locking and unlocking can
fail. Thus the signature was changed to include error returns.

Note that this error is currently not handled well in this POC, however
an action item to replace the makeshift error handling with proper error
handling has been created.

Another major change is that we no longer use one lock, that is passed
through the entire app by reference. This is because the etcd RWMutex
works fundamentally different. The "shared" stated will be kept
server-side, this means locally the type should only be shared from a
single Lock to an Unlock. This is not yet documented well on the code
and will be added in an upcoming commit.