gh-685 manually extract context in handlers
there was a breaking change in go-swagger. In v16 you could run the
generation with --with-context and context would always be the first
argument. With v18 that option no longer exists. However, we can still
manually extract the context from the underlying HTTPRequest which we
can get throught the params.