gh-645: extract schemaC11y out of regular c11y, provide as source to gQL
Previously we added the new schema-related methods directly into the
core contextionary. However, that made the contextionary very specific
to weaviate's causes. Now that we've moved this to the other
(schema-related) contextionary folder, we have decoupled this a bit.

Additionally, the c11y is now available to the the graphQL API tree (as
part of the root map).