gh-761: bugfix: use custom gql fragments on cref unions
We previously assumed that the object inside the Class property
SelectionSet would either be a Field or an InlineFragment. However, if
the user uses custom fragments, we can also run into FragmentSpreads. We
are now able to deal with them as well (and recursively too)

Note that one part isn't particular pretty due to:
https://github.com/graphql-go/graphql/issues/455